[#57] refactor - 텅 빔 페이지 모듈화
